+++ msm6295.h
@@ -2,7 +2,7 @@
 
 #define MAX_MSM6295 (2)
 
-int MSM6295Init(int nChip, int nSamplerate, float fMaxVolume, char bAddSignal);
+int MSM6295Init(int nChip, int nSamplerate, double fMaxVolume, char bAddSignal);
 void MSM6295Reset(int nChip);
 void MSM6295Exit(int nChip);
